requirements.txt
flask
flask-socketio
python-dotenv
requests
google-generativeai
anthropic
openai
